Add core sliding sync classes
